Add files via upload
My proposal is that we will prioritize testing who has to be tested, it will be prioritized if we test the right person at the right time then we can reduce the infection of the virus. There is a layer of technology on the top of the health cate system, which tells the health care system which one to test first.